Add global and scene stylesheets
The global stylesheet is determined by the existence on the gist of a file
named style.css. It get applied during the initialization of the story
before the first scene is rendered.

Scene's stylesheet is associated to the `style` property of the YAML Front
Matter. The stylesheet is enabled when the scene is active and disabled
when the scene is not active.